+++ extensions/data-plane-transfer/data-plane-transfer-sync/src/main/java/org/eclipse/dataspaceconnector/transfer/dataplane/sync/proxy/DataPlaneTransferProxyReferenceServiceImpl.java
@@ -1,0 +1,71 @@
+/*
+ *  Copyright (c) 2022 Amadeus
+ *
+ *  This program and the accompanying materials are made available under the
+ *  terms of the Apache License, Version 2.0 which is available at
+ *  https://www.apache.org/licenses/LICENSE-2.0
+ *
+ *  SPDX-License-Identifier: Apache-2.0
+ *
+ *  Contributors:
+ *       Amadeus - initial API and implementation
+ *
+ */
+
+package org.eclipse.dataspaceconnector.transfer.dataplane.sync.proxy;
+
+import org.eclipse.dataspaceconnector.common.token.TokenGenerationService;
+import org.eclipse.dataspaceconnector.dataplane.spi.DataPlaneConstants;
+import org.eclipse.dataspaceconnector.spi.result.Result;
+import org.eclipse.dataspaceconnector.spi.types.TypeManager;
+import org.eclipse.dataspaceconnector.spi.types.domain.edr.EndpointDataReference;
+import org.eclipse.dataspaceconnector.transfer.dataplane.spi.proxy.DataPlaneTransferProxyCreationRequest;
+import org.eclipse.dataspaceconnector.transfer.dataplane.spi.proxy.DataPlaneTransferProxyReferenceService;
+import org.eclipse.dataspaceconnector.transfer.dataplane.spi.security.DataEncrypter;
+import org.jetbrains.annotations.NotNull;
+
+import java.time.Instant;
+import java.util.Date;
+import java.util.HashMap;
+
+import static org.eclipse.dataspaceconnector.dataplane.spi.DataPlaneConstants.CONTRACT_ID;
+
+public class DataPlaneTransferProxyReferenceServiceImpl implements DataPlaneTransferProxyReferenceService {
+
+    private final TokenGenerationService tokenGenerationService;
+    private final TypeManager typeManager;
+    private final long tokenValiditySeconds;
+    private final DataEncrypter dataEncrypter;
+
+    public DataPlaneTransferProxyReferenceServiceImpl(TokenGenerationService tokenGenerationService, TypeManager typeManager, long tokenValiditySeconds, DataEncrypter dataEncrypter) {
+        this.tokenGenerationService = tokenGenerationService;
+        this.typeManager = typeManager;
+        this.tokenValiditySeconds = tokenValiditySeconds;
+        this.dataEncrypter = dataEncrypter;
+    }
+
+    /**
+     * Creates an {@link EndpointDataReference} targeting public API of the provided Data Plane so that it is used
+     * as a proxy to query the data from the data source.
+     */
+    @Override
+    public Result<EndpointDataReference> createProxyReference(@NotNull DataPlaneTransferProxyCreationRequest request) {
+        var encryptedDataAddress = dataEncrypter.encrypt(typeManager.writeValueAsString(request.getContentAddress()));
+        var decorator = new DataPlaneProxyTokenDecorator(Date.from(Instant.now().plusSeconds(tokenValiditySeconds)), request.getContractId(), encryptedDataAddress);
+        var tokenGenerationResult = tokenGenerationService.generate(decorator);
+        if (tokenGenerationResult.failed()) {
+            return Result.failure(tokenGenerationResult.getFailureMessages());
+        }
+
+        var props = new HashMap<>(request.getProperties());
+        props.put(CONTRACT_ID, request.getContractId());
+
+        var builder = EndpointDataReference.Builder.newInstance()
+                .id(request.getId())
+                .endpoint(request.getProxyEndpoint())
+                .authKey(DataPlaneConstants.PUBLIC_API_AUTH_HEADER)
+                .authCode(tokenGenerationResult.getContent().getToken())
+                .properties(props);
+        return Result.success(builder.build());
+    }
+}